+++ CHANGELOG.md
@@ -5,7 +5,7 @@ file.
 
 The project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## v1.0.0 (2025-12-18)
 
 ### Added